Final Commit
Final commit of the project. Alien Spore Creature model is committed.
.usdc file was too big to commit and use repo after. Committing this file uses all the space on the repo.
A series of issues with committing this file has resulted in me uploading it to Github.com instead of desktop. The file won't be in the correct place for the project. This file had to be reduced in size to upload. It has every elements except the model removed to reduce file size.
This final commit was discussed and agreed in the Beta presentation to be committed at a later date after the beta submission.